# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.72.65.211"
  description: SwaggerHub API Auto Mocking hey
tags:
- name: Auth
  description: All authentication related routes
- name: Users
  description: All the user's personal app info
- name: Recipes
  description: All reciepes general info
paths:
  /register:
    post:
      tags:
      - Auth
      summary: A new user from the client is added to the server
      description: A JSON object contains all the details about the new user.
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
  /login:
    post:
      tags:
      - Auth
      summary: Login an exist user
      description: An exist user login to the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginScheme'
        required: true
      responses:
        "200":
          description: |
            User was found in the DB and authenitcated successfully. The session ID is returned in a cookie named `session`.  You should include this cookie in subsequent requests
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: Authentication failed
  /logout:
    post:
      tags:
        - Auth
      summary: logout a user from the system and delete his cookie
      responses:
        "200":
          description: "the user has ben successfully logged out"
  /users/removeFavorite:
    post:
      tags:
      - Users
      summary: remove recipe in the favorites list of a user
      description: remove recipe in the favorites list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user removed
                  example: 716429
      responses:
        "200":
          description: The Recipe successfully removed
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/favorites:
    post:
      tags:
      - Users
      summary: save recipe in the favorites list of a user
      description: save recipe in the favorites list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user saved
                  example: 716429
      responses:
        "200":
          description: The Recipe successfully saved as favorite
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
    get:
      tags:
      - Users
      summary: Return the favorite recipes that were saved by user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/userOn:
    get:
      tags:
      - Users
      summary: Return the last three recipes that user view
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's favorites recipes returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipePreview'
                  favorites:
                    $ref: '#/components/schemas/Favorite'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/myFamily:
    post:
      tags:
      - Users
      summary: save recipe in the family list of a user
      description: save recipe in the family list of a user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reciepeId:
                  type: string
                  description: reciepe's id which the user saved
                  example: 2
      responses:
        "200":
          description: The Recipe successfully saved as viewed
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
    get:
      tags:
      - Users
      summary: Return the family recipes that user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's family recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /users/myRecipe:
    get:
      tags:
      - Users
      summary: Return the recipes of the user
      description: Get an array of recipes that were saved by a specific user
      responses:
        "200":
          description: data about the user's recipes returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        "401":
          description: user is not authorized
      security:
      - cookieAuth: []
  /recipes/showRecipe:
    get:
      tags:
      - Recipes
      summary: Return the missing details on recipe by his ID
      description: Return the missing details on recipe by his ID
      parameters:
      - name: recipe_id
        in: query
        description: recipe's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: data about the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Show'
        "404":
          description: recipe_id is not exists
  /recipes/randomRecipes/:
    get:
      tags:
      - Recipes
      summary: Return three random recipes
      description: Get an array of recipes
      responses:
        "200":
          description: data about three random recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipePreview'
                  favorites:
                    $ref: '#/components/schemas/Favorite'
                  views:
                    $ref: '#/components/schemas/View'
  /recipes/search/:
    get:
      tags:
      - Recipes
      summary: Return recipes according to the search
      description: Get an array of recipes
      parameters:
      - name: numberOfRecipe
        in: query
        description: how much recipe to return
        style: simple
        explode: false
        required: true
        schema:
          type: integer
      - name: RecipeName
        in: query
        description: recipe's name
        style: simple
        explode: false
        schema:
          type: string
      - name: cuisine
        in: query
        description: filter according cuisine
        style: simple
        explode: false
        schema:
          type: string
      - name: diet
        in: query
        description: filter according diet
        style: simple
        explode: false
        schema:
          type: string
      - name: intolerance
        in: query
        description: filter according intolerance
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: data about three random recipes
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipePreview'
                  favorites:
                    $ref: '#/components/schemas/Favorite'
                  views:
                    $ref: '#/components/schemas/View'
  /recipes/createRecipe/:
    post:
      tags:
      - Recipes
      summary: A new recipe from the client is added to the server
      description: A JSON object contains all the details about the new user.
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipe'
        required: true
      responses:
        "201":
          description: A new recipe has been added
          content:
            application/json:
              schema:
                type: object
                properties:
                  msessage:
                    type: string
                    example: Recipe 1 was created successfully
                  success:
                    type: boolean
                    example: true
                  recipe_id:
                    type: integer
                    example: 1
        "400":
          description: wrong input parameters
components:
  schemas:
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: my_name
        firstname:
          type: string
          description: user's first name
          example: my_first_name
        lastname:
          type: string
          description: user's last name
          example: my_last_name
        country:
          type: string
          description: user's country
          example: Israel
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: my_name@12
        email:
          pattern: email
          type: string
          description: user's email
          example: my_name12@gmail.com
        # image-url:
        #   pattern: URL
        #   type: string
        #   description: URL for the user's image
        #   example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins letters only
        password:
          maxLength: 10
          minLength: 5
          description: password conatins at least one number and one special char
          format: password
      example:
        username: my_name
        password: my_name@12
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 45
        popularity:
          type: integer
          description: how many likes got this recipe
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
    Favorite:
      type: object
      additionalProperties:
        type: boolean
    Recipe:
      type: object
      required:
        - recipeName
      properties:
        recipeName:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        preparationTile:
          type: string
          description: the time takes to make this recipe
          example: 40 min
        clickable:
          type: boolean
          description: if the recipe is clickale
          example: true
        ingredients:
          type: string
          description: ingredients of the recipe
          example: sugar:50g, milk:10ml
        preparationInstructions:
          type: string
          description: Insructions of the recipe
          example: Add 50 gram of sugar to the bowl
        numberOfDishes:
          type: integer
          description: number of dishes 
          example: 4
        vegetarianOrVegan:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        gluten:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
        owner:
          type: string
          description: The owner of the recipe from the family
          example: my grandma
        customaryPrepare:
          type: string
          description: when the recipe prepared
          example: friday night
    Show:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 1
        extendedIngredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredients'
          description: ingredients of the recipe
        instructions:
          type: string
          description: instructions of the recipe
        servings:
          type: integer
          description: number of dishes
    Ingredients:
      type: object
      properties:
        name:
          type: string
          description: name of recipe
          example: fresh basil
        amount:
          type: object
          properties:
            us:
              type: object
              properties:
                amount:
                  type: number
                  example: 0.25
                unsitShort:
                  type: string
                  example: cups
                unitLong:
                  type: ml
                  example: milliliters
            metric:
              type: object
              properties:
                amount:
                  type: number
                  example: 0.25
                unsitShort:
                  type: string
                  example: cups
                unitLong:
                  type: ml
                  example: milliliters
    CreateRecipe:
      type: object
      properties:
        recipeName:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        preparationTile:
          type: string
          description: the time takes to make this recipe
          example: 40 min
        clickable:
          type: boolean
          description: if the recipe is clickale
          example: true
        ingredients:
          type: string
          description: ingredients of the recipe
          example: sugar:50g, milk:10ml
        preparationInstructions:
          type: string
          description: Insructions of the recipe
          example: Add 50 gram of sugar to the bowl
        numberOfDishes:
          type: integer
          description: number of dishes 
          example: 4
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        Vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false  
        gluten:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
        owner:
          type: string
          description: The owner of the recipe from the family
          example: my grandma
        customaryPrepare:
          type: string
          description: when the recipe prepared
          example: friday night
    View:
      type: object
      additionalProperties:
        type: boolean
  